src/rest_function/requirements.txt
azure-functions>=1.24.0
azure-storage-blob>=12.20.0
azure-identity>=1.16.0
